Changed to enable multiple project files
Updated the DacVersion activity so that it has a Files property like the
TfsVersion activity so that it's possible to have more than one file
versioned at the same time. I left the current property intact so as not
to break working builds. This new property also enables using the
FindMatchingFiles activity to find the files to act upon and so create a
more dynamic build workflow.

Also changed how the activity finds the DacVersion tag to use
XmlDocument class instead of naive string search that does not account
for any version of the tag other than opening and closing tags being on
the same line.

Added a new unit test to test the new functionality and fixed the old
test so that it does not rely on an external file that is not a part of
the solution in a specific place to work, so now it works the same as
the TfsVersion activity tests.