Changing the character encoding for Sonar
SonarQube uses the Java Properties API to read configuration files. This API uses the 'Latin alphabet no. 1' character set for its input and output. With this patch applied, the Sonar activity uses the exact same character set for reading and writing configuration files.

Previously, the Sonar activity used the ASCII set, which is only a subset of all the characters that SonarQube supports (ISO 8859-1).

Fixes #7